rename beb to eventmesh in pkg/sender/
